バーチャルパットとキャラクター画像の連携 2
以下変更内容
クラス名の変更
ariableJoystickをFloatingJoystickに変更。
