+++ 2DActionGame/Assets/Adaptive Performance.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: aca7d68b33fc767409fa511eefa40ad6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
