+++ 2DActionGame/Assets/Adaptive Performance/Settings.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 92ca7a8eb1aa2f54fb3d4884409cac21
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
