+++ 2DActionGame/Assets/Adaptive Performance/Settings/Simulator Provider Settings.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6a6c37ec727324944b0cc4adfdf7f9c6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
