+++ 2DActionGame/Assets/script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e9763cb226b0554d9f140f30d99567e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
