+++ 2DActionGame/ProjectSettings/EditorBuildSettings.asset
@@ -8,4 +8,5 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/SampleScene.unity
     guid: 2cda990e2423bbf4892e6590ba056729
-  m_configObjects: {}
+  m_configObjects:
+    com.unity.adaptiveperformance.simulator.provider_settings: {fileID: 11400000, guid: 6a6c37ec727324944b0cc4adfdf7f9c6, type: 2}
